enable autofold again but only for imports
The reason why I'm exluding comments is because there is a bug in the
typescript language server.
When you have a class that contains a comment, and close that fold, then
use `vim.lsp.foldclose("comment", winid)` to close all comment folds,
the class fold is closed, which should not happen.

Example Class:

```typescript
class Test {
    /*
     * Test
     */
}
```